% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{dalys}
\alias{dalys}
\title{Disability adjusted life years}
\usage{
dalys(
  x,
  age_divisor,
  mild_disability_weight = 0.006,
  moderate_disability_weight = 0.051,
  severe_disability_weight = 0.133,
  clinical_episode_length = 0.01375 * 365,
  severe_episode_length = 0.04795 * 365
)
}
\arguments{
\item{x}{Input data.frame}

\item{age_divisor}{To transform the age units. For example setting 365 will return age units in years for age-disaggregated outputs}

\item{mild_disability_weight}{disability weight for mild malaria. Assigned to clinical cases in those over 5 years old}

\item{moderate_disability_weight}{disability weight for moderate malaria. Assigned to clinical cases in those under 5 years old}

\item{severe_disability_weight}{disability weight for severe malaria. Assigned to all severe cases}

\item{clinical_episode_length}{average length of an episode of clinical malaria}

\item{severe_episode_length}{average length of an episode of severe malaria}
}
\description{
Calculate Years Lived with Disability (YLDs) and Disability-Adjusted Life-Years
based on disability weights from the Global Burden of Disease study. To estimate
YLL we assume the average life expectancy for a person aged x years taken from the UN WPP
Africa profile for 2022. To estimate the expected age in a given range we assume
exponentially distributed ages with the range. 
YLLs are calculated by multiplying deaths by the number of years an individual was expected to live past their year of death.
}
\details{
Disability weights sourced \href{https://ghdx.healthdata.org/record/ihme-data/gbd-2017-disability-weights}{here}

Duration of disease was sourced from \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5321383/}{Winskill et al 2017 (Supplementary Table 8)}

This is an approximation of YLD estimation from the GBD study; disability due to comorbid conditions
such as motor impairment and anemia are excluded.

Note: weights and lengths are currently for Plasmodium falciparum.
}
