% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{drop_burnin}
\alias{drop_burnin}
\title{Remove burn in period from simulation output}
\usage{
drop_burnin(x, burnin)
}
\arguments{
\item{x}{Input data.frame}

\item{burnin}{Length of burn in period in days}
}
\description{
Remove burn in period from simulation output
}
