% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{format_time}
\alias{format_time}
\title{Adds time columns}
\usage{
format_time(x, baseline_year)
}
\arguments{
\item{x}{Input data.frame}

\item{baseline_year}{baseline_year Baseline year}
}
\description{
Adds year (assuming no leap years), month, week (assuming last week in the year is 8 days),
day and time columns. Time is usuful for plotting.
}
