% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{prevalence_aggregate}
\alias{prevalence_aggregate}
\title{Aggregate prevalence over time}
\usage{
prevalence_aggregate(x, ...)
}
\arguments{
\item{x}{Input data.frame}

\item{...}{Aggregation columns}
}
\description{
Aggregate prevalence over time
}
