% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{prevalence_format}
\alias{prevalence_format}
\title{Format prevalence output}
\usage{
prevalence_format(x, age_divisor = 365)
}
\arguments{
\item{x}{Input datas.frame}

\item{age_divisor}{Aggregation level. For example setting to 365 will return
age units in years}
}
\description{
Format prevalence output
}
