% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{prevalence_time_aggregate}
\alias{prevalence_time_aggregate}
\title{Aggregate prevalence over time}
\usage{
prevalence_time_aggregate(x)
}
\arguments{
\item{x}{Input data.frame}
}
\description{
Aggregate prevalence over time
}
