% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{rates_age_aggregate}
\alias{rates_age_aggregate}
\title{Aggregate rates output over age}
\usage{
rates_age_aggregate(x)
}
\arguments{
\item{x}{Input data.frame}
}
\description{
Aggregate rates output over age
}
