% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{rates_aggregate}
\alias{rates_aggregate}
\title{Aggregate rates}
\usage{
rates_aggregate(rates, ...)
}
\arguments{
\item{rates}{formatted rates}

\item{...}{Aggregation columns}
}
\description{
Aggregate rates
}
