% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{rates_column_check}
\alias{rates_column_check}
\title{Rates input checks}
\usage{
rates_column_check(x)
}
\arguments{
\item{x}{Input data.frame}
}
\description{
Checks that the required columns are present and that the age-ranges for
required columns align.
}
