% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{rates_time_aggregate}
\alias{rates_time_aggregate}
\title{Aggregate rates output over t}
\usage{
rates_time_aggregate(x)
}
\arguments{
\item{x}{Input data.frame}
}
\description{
Aggregate rates output over t
}
