% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{severe_incidence_mortality}
\alias{severe_incidence_mortality}
\title{Adjust severe rates as a result of treatment coverage, dis-aggregated by hospitalised and community and estimate mortality}
\usage{
severe_incidence_mortality(
  x,
  treatment_scaler,
  hosp_sev_cfr,
  community_sev_cfr
)
}
\arguments{
\item{x}{Input data.frame}

\item{treatment_scaler}{Proportion of treated severe cases averted by treatment}

\item{hosp_sev_cfr}{Severe case fatality ratio in hospital
The original estimate fitted to Reyburn et al data in Griffin et al (2016) 0.065}

\item{community_sev_cfr}{Severe case fatality ratio in the commnity (non-hospitalised)
The original estimate fitted to Lubell et al data in Griffin et al (2016) 0.6}
}
\description{
See\href{https://www.thelancet.com/journals/laninf/article/PIIS1473-3099(15)00423-5/fulltext}{Griffin et al (2016)} SI for
more details
}
