% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{time_transform}
\alias{time_transform}
\title{Create a new time column for aggregation}
\usage{
time_transform(x, time_divisor = 1, baseline_t = 0)
}
\arguments{
\item{x}{Input data.frame}

\item{time_divisor}{Aggregation level. Default = 1 (no aggregation).
Setting to 365 would allow annual aggregation, 30 monthly, 7 weekly etc.}

\item{baseline_t}{A baseline time to add to the time output}
}
\description{
Create a new time column for aggregation
}
