% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{treatment_scaling}
\alias{treatment_scaling}
\title{Adjust severe (and downstream mortality) rates as a result of treatment coverage.}
\usage{
treatment_scaling(x, treatment_scaler, baseline_treatment = 0)
}
\arguments{
\item{x}{Input data.frame}

\item{treatment_scaler}{The impact of treatment coverage on progression to severe disease and death.
The probability of being hospitalised for untreated cases compared to treated cases is difficult to study and is not known, but can 
be estimated from available data with several assumptions. Our estimate is derived from data used in a meta-analysis on the impact 
of delayed treatment of uncomplicated malaria on progression to severe malaria 
\href{https://journals.plos.org/plosmedicine/article?id=10.1371/journal.pmed.1003359}{(Mousa et al, 2020, Plos Medicine)}, which estimated 
the relative risk ratio for treatment delays of 1-2, 2-3 and >3 days compared to treatment within 1 day. Here, we assume that the 
risk of hospitalisation with severe malaria among untreated cases would be double the risk in this last “delay category” of >3 days 
(2 * RR 1.186), giving a relative risk of any severe malaria of 0.578 for treatment within 1 day compared to no treatment (i.e. a treatment
scaler of 0.42 for the proportion of severe malaria cases that are prevented by treatment of uncomplicated malaria). However, in sensitivity
analysis a range of 0.281-0.843 should be considered for the treatment scaler, which was derived by assuming the risk among untreated cases
would be the same as the risk with a treatment delay of >3 days (lower estimate) or 3 times the risk as those with a treatment delay of
>3 days (upper estimate).}

\item{baseline_treatment}{The proportion of uncomplicated malaria cases that are effectively treated historically.}
}
\description{
Adjust severe (and downstream mortality) rates as a result of treatment coverage.
}
