% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{ea}
\alias{ea}
\title{Expected age within range assuming exponentially distributed pop}
\usage{
ea(lower_age, upper_age, average_age)
}
\arguments{
\item{lower_age}{Lower bound of age range (single value)}

\item{upper_age}{Upper bound of age range (single value)}

\item{average_age}{Average age of population}
}
\description{
Expected age within range assuming exponentially distributed pop
}
