% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R, R/dalys.R
\name{ea}
\alias{ea}
\title{Expected age within range assuming exponentially distributed pop}
\usage{
ea(age_lower, age_upper, average_age)

ea(age_lower, age_upper, average_age)
}
\arguments{
\item{age_lower}{Lower bound of age range (single value)}

\item{age_upper}{Upper bound of age range (single value)}

\item{average_age}{Average age of population}
}
\description{
Expected age within range assuming exponentially distributed pop

Expected age within range assuming exponentially distributed pop
}
