% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R, R/dalys.R
\name{expected_age}
\alias{expected_age}
\title{Expected age within range assuming exponentially distributed pop}
\usage{
expected_age(age_lower, age_upper, average_age)

expected_age(age_lower, age_upper, average_age)
}
\arguments{
\item{age_lower}{Lower bound of age range (can be a vector)}

\item{age_upper}{Upper bound of age range (can be a vector)}

\item{average_age}{Average age of population}
}
\description{
Expected age within range assuming exponentially distributed pop

Expected age within range assuming exponentially distributed pop
}
