% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{expected_age}
\alias{expected_age}
\title{Expected age within range assuming exponentially distributed pop}
\usage{
expected_age(lower_age, upper_age, average_age)
}
\arguments{
\item{lower_age}{Lower bound of age range (can be a vector)}

\item{upper_age}{Upper bound of age range (can be a vector)}

\item{average_age}{Average age of population}
}
\description{
Expected age within range assuming exponentially distributed pop
}
