% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{get_prevalence}
\alias{get_prevalence}
\title{Extract basic rates from model output}
\usage{
get_prevalence(x, time_divisor, baseline_t, age_divisor)
}
\arguments{
\item{x}{Input data.frame}

\item{age_divisor}{Aggregation level. For example setting to 365 will return
age units in years}
}
\description{
Extract basic rates from model output
}
