% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{get_rates}
\alias{get_rates}
\title{Extract basic rates from model output}
\usage{
get_rates(
  x,
  baseline_year = 2000,
  ages_as_years = TRUE,
  scaler = 0.215,
  treatment_scaler = 0.5
)
}
\arguments{
\item{x}{Input data.frame}

\item{ages_as_years}{Converts ages from days to years}

\item{scaler}{Scaler for severe cases to deaths. The
original fitted scaler for P. falciparum is 0.215. See
\href{https://www.thelancet.com/journals/laninf/article/PIIS1473-3099(15)00423-5/fulltext}{Griffin et al (2016)}.}

\item{treatment_scaler}{The impact of treatment coverage on progression to severe disease and death.
A highly uncertain parameter, the analysis by
 \href{https://www.thelancet.com/journals/laninf/article/PIIS1473-3099(15)00423-5/fulltext}{Griffin et al (2016)}
sampled from a uniform diastribution (0, 1).}

\item{baseline}{year Starting year, assuming the model runs begins at the start of the year}
}
\description{
Extract basic rates from model output
}
