% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{get_rates}
\alias{get_rates}
\title{Extract basic rates from malariasimulation model output}
\usage{
get_rates(
  x,
  time_divisor,
  baseline_t,
  age_divisor,
  scaler = 0.215,
  treatment_scaler = 0.42,
  aggregate_age = FALSE,
  mild_disability_weight = 0.006,
  moderate_disability_weight = 0.051,
  severe_disability_weight = 0.133,
  clinical_episode_length = 0.01375,
  severe_episode_length = 0.04795
)
}
\arguments{
\item{x}{Input data.frame}

\item{time_divisor}{Aggregation level. Default = 1 (no aggregation).
Setting to 365 would allow annual aggregation, 30 monthly, 7 weekly etc.}

\item{baseline_t}{A baseline time to add to the time output}

\item{age_divisor}{To transform the age units. For example setting 365 will return age units in years for age-disaggregated outputs}

\item{scaler}{Scaler for severe cases to deaths. The
original fitted scaler for P. falciparum is 0.215. See
\href{https://www.thelancet.com/journals/laninf/article/PIIS1473-3099(15)00423-5/fulltext}{Griffin et al (2016)}.}

\item{treatment_scaler}{The impact of treatment coverage on progression to severe disease and death.
The probability of being hospitalised for untreated cases compared to treated cases is difficult to study and is not known, but can 
be estimated from available data with several assumptions. Our estimate is derived from data used in a meta-analysis on the impact 
of delayed treatment of uncomplicated malaria on progression to severe malaria 
\href{https://journals.plos.org/plosmedicine/article?id=10.1371/journal.pmed.1003359}{(Mousa et al, 2020, Plos Medicine)}, which estimated 
the relative risk ratio for treatment delays of 1-2, 2-3 and >3 days compared to treatment within 1 day. Here, we assume that the 
risk of hospitalisation with severe malaria among untreated cases would be double the risk in this last “delay category” of >3 days 
(2 * RR 1.186), giving a relative risk of any severe malaria of 0.578 for treatment within 1 day compared to no treatment (i.e. a treatment
scaler of 0.42 for the proportion of severe malaria cases that are prevented by treatment of uncomplicated malaria). However, in sensitivity
analysis a range of 0.281-0.843 should be considered for the treatment scaler, which was derived by assuming the risk among untreated cases
would be the same as the risk with a treatment delay of >3 days (lower estimate) or 3 times the risk as those with a treatment delay of
>3 days (upper estimate).}

\item{aggregate_age}{Aggregate output over age groups}

\item{mild_disability_weight}{disability weight for mild malaria. Assigned to clinical cases in those over 5 years old}

\item{moderate_disability_weight}{disability weight for moderate malaria. Assigned to clinical cases in those under 5 years old}

\item{severe_disability_weight}{disability weight for severe malaria. Assigned to all severe cases}

\item{clinical_episode_length}{average length of an episode of clinical malaria}

\item{severe_episode_length}{average length of an episode of severe malaria}
}
\description{
Summarises clinical incidence `clinical`, severe incidence `severe` and mortality `mortality`
rates as well as per-capita years lived with disability `yld_pp`, years of life
lost `yll_pp` and daly `dalys_pp`.
}
\details{
For DALYs: disability weights are from the Global Burden of Disease study. To estimate
YLL we assume the average life expectancy for a person aged x years taken from the UN WPP
Africa profile for 2022. To estimate the expected age in a given range we assume
exponentially distributed ages with the range. Disability weights sourced:
 \href{https://ghdx.healthdata.org/record/ihme-data/gbd-2017-disability-weights}{here}.
This is an approximation of YLD estimation from the GBD study; disability due to comorbid conditions
such as motor impairment and aneamia are excluded.

Note: Default parameter values are for _Plasmodium falciparum_
}
