% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{prevalence_format}
\alias{prevalence_format}
\title{Format prevalence output}
\usage{
prevalence_format(x, ages_as_years)
}
\arguments{
\item{x}{Input data.frame}

\item{ages_as_years}{Convert ages to be in units of years}
}
\description{
Format prevalence output
}
