% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{rates_format}
\alias{rates_format}
\title{Format rates}
\usage{
rates_format(x, clinical_cols, severe_cols, ages_as_years)
}
\arguments{
\item{x}{Input data.frame}

\item{clinical_cols}{Clinical incidence column names}

\item{severe_cols}{Severe incidence column names}

\item{age_divisor}{Aggregation level. For example setting to 365 will return
age units in years}
}
\description{
Create rates from counts and specifies age output units
}
