% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{treatment_scaling}
\alias{treatment_scaling}
\title{Adjust severe (and downstream mortality) rates as a result of treatment coverage.}
\usage{
treatment_scaling(x, treatment_scaler, baseline_treatment = 0)
}
\arguments{
\item{x}{Input data.frame}

\item{treatment_scaler}{The impact of treatment coverage on progression to severe disease and death.
A highly uncertain parameter, the analysis by
 \href{https://www.thelancet.com/journals/laninf/article/PIIS1473-3099(15)00423-5/fulltext}{Griffin et al (2016)}
sampled from a uniform diastribution (0, 1).}

\item{baseline_treatment}{The proportion of uncomplicated malaria cases that are effectively treated historically.}
}
\description{
Adjust severe (and downstream mortality) rates as a result of treatment coverage.
}
